+++ R/vistime.R
@@ -114,12 +114,15 @@
 #' pp
 #' }
 #'
-vistime <- function(data, events = "event", start = "start", end = "end", groups = "group", colors = "color", fontcolors = "fontcolor", tooltips = "tooltip", linewidth = NULL, title = NULL, showLabels = NULL, show_labels = TRUE, lineInterval = NULL, background_lines = 10) {
+vistime <- function(data, events = "event", start = "start", end = "end", groups = "group",
+                    colors = "color", fontcolors = "fontcolor", tooltips = "tooltip",
+                    optimize_y = TRUE, linewidth = NULL, title = NULL, showLabels = NULL,
+                    show_labels = TRUE, lineInterval = NULL, background_lines = 10) {
   data <- validate_input(data, start, end, events, groups, linewidth, title, showLabels, show_labels, lineInterval, background_lines)
   data <- set_colors(data, colors, fontcolors)
   data <- fix_columns(data, events, start, end, groups, tooltips)
   data <- set_subplots(data)
-  data <- set_y_values(data)
+  data <- set_y_values(data, optimize_y)
 
   ranges <- plot_ranges(data, linewidth, show_labels, background_lines)
   events <- plot_events(data, show_labels, background_lines)