+++ R/plot_events.R
@@ -35,18 +35,18 @@ plot_events <- function(data_orig, show_labels, background_lines) {
     thisData$labelY <- thisData$y + 0.5 * rep_len(c(-1, 1), nrow(thisData))
 
     # add vertical lines to plot
-    p <- plot_ly(thisData, type = "scatter", mode = "markers")
+    p <- plotly::plot_ly(thisData, type = "scatter", mode = "markers")
 
     # 1. add vertical line for each year/day
     for (day in seq(min(data_orig$start), max(data_orig$end), length.out = background_lines + 1)) {
-      p <- add_lines(p,
+      p <- plotly::add_lines(p,
         x = as.POSIXct(day, origin = "1970-01-01"), y = c(0, maxY),
-        line = list(color = toRGB("grey90")), showlegend = F, hoverinfo = "none"
+        line = list(color = plotly::toRGB("grey90")), showlegend = F, hoverinfo = "none"
       )
     }
 
     # add all the markers for this Category
-    p <- add_markers(p,
+    p <- plotly::add_markers(p,
       x = ~start, y = ~y,
       marker = list(
         color = ~col, size = 15, symbol = "circle",
@@ -57,14 +57,14 @@ plot_events <- function(data_orig, show_labels, background_lines) {
 
     # add annotations or not
     if (show_labels) {
-      p <- add_text(p,
-        x = ~start, y = ~labelY, textfont = list(family = "Arial", size = 14, color = ~ toRGB(fontcol)),
+      p <- plotly::add_text(p,
+        x = ~start, y = ~labelY, textfont = list(family = "Arial", size = 14, color = ~ plotly::toRGB(fontcol)),
         textposition = ~labelPos, showlegend = F, text = ~label, hoverinfo = "none"
       )
     }
 
     # fix layout
-    p <- layout(p,
+    p <- plotly::layout(p,
       hovermode = "closest",
       xaxis = list(showgrid = F, title = ""),
       yaxis = list(
@@ -78,4 +78,3 @@ plot_events <- function(data_orig, show_labels, background_lines) {
   names(events) <- eventNumbers # preserve order of subplots
   return(events)
 }
-#' @noRd