+++ tests/testthat/test_set_y_values.R
@@ -21,7 +21,7 @@ dat <- data.frame(
   end = c("2019-01-01", "2019-01-10")
 )
 
-dat <- vistime:::validate_input(dat, start, end, events, groups, linewidth, title, showLabels, show_labels, lineInterval, background_lines)
+dat <- vistime:::validate_input(dat, start, end, events, groups, tooltips, optimize_y, linewidth, title, showLabels, show_labels, lineInterval, background_lines)
 dat <- vistime:::set_colors(dat, colors, fontcolors)
 dat <- vistime:::fix_columns(dat, events, start, end, groups, tooltips)
 dat <- vistime:::set_subplots(dat)