+++ NEWS.md
@@ -1,5 +1,8 @@
 # 0.8.1.9000 
 - fixed bug for plots having more than 9 groups - in these cases the order was not the same as in the input data frame
+- new argument `optimize_y` (default: `TRUE`)
+  * if `optimize_y = TRUE`, use heuristic to optimally distribute events on y axis
+  * if `optimize_y = FALSE`, use fixed order from input data
 
 # 0.8.1
 - hotfix: colors and fontcolors handling was broken.